ci: T-044 workspace green-gate (phase-2 CI R8 / build-site close-out)
Tier 8 / tests R8 / final Phase 2 task. The existing ci.yml had drifted —
clippy version bumps + organic fmt drift meant `-D warnings` was silently
permissive on PRs. This commit restores the gate honestly:

- .github/workflows/ci.yml
  * clippy step now runs with --locked + a documented `-A` allowlist
    (14 lints, each deliberate — figment::Error size, fmt-triggered
    doc-comment re-indents, etc.). Rationale in workflow header.
  * unit build step bumped to `--workspace --all-targets --locked`.
  * unit test step pinned to `--tests` per Phase 2 R8 acceptance
    ("no features, no env vars; integration + trybuild MUST run").
- 19 small clippy-fix edits across 15 crates (io_other_error,
  unwrap_or_default, manual_contains, inspect_err, sort_by_key,
  vec_init_then_push, single_match, assert-eq-literal-bool,
  match_like_matches_macro, …). No API changes.
- 5 `#[allow(dead_code)]` for legit test-only helpers (list::render_table,
  list::cwd_for, kill::ulid_a/b, ark-view marker-impl test structs).
- Ledger entries (impl-soul-phase-2.md + loop-log.md) prepended.

Local validation: fmt ✓ · clippy ✓ (0 errors) · build ✓ · test ✓
(2057 pass / 4 ignored / 0 fail / 62 suites / 17.84s).

Phase 2 CLOSED: 45/45 tasks. Coverage 36/36 R's.

+++ crates/plugins/picker/src/bootstrap.rs
@@ -369,7 +369,7 @@ fn days_from_civil(y: i64, m: u32, d: u32) -> Option<i64> {
     }
     let y = if m <= 2 { y - 1 } else { y };
     let era = y.div_euclid(400);
-    let yoe = (y - era * 400) as i64; // [0, 399]
+    let yoe = y - era * 400; // [0, 399]
     let m_adj = if m > 2 { m as i64 - 3 } else { m as i64 + 9 };
     let doy = (153 * m_adj + 2) / 5 + d as i64 - 1; // [0, 365]
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy; // [0, 146096]
